@@ -5,7 +5,8 @@ var list_view_1 = require("ui/list-view");
 var viewsModule = require("./shared/views");
 var navigationModule = require("navigation");
 var constants_1 = require("./shared/constants");
-var chartModule = require("nativescript-telerik-ui/chart");
+var chartModule = require("nativescript-telerik-ui-pro/chart");
+var listViewModule = require("nativescript-telerik-ui-pro/listview");
 var months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
 specialPropertiesModule.registerSpecialProperty("link", function (instance, propertyValue) {
     if (instance instanceof list_view_1.ListView) {
@@ -13,6 +14,14 @@ specialPropertiesModule.registerSpecialProperty("link", function (instance, prop
         listView.on("itemTap", function (args) {
             navigationModule.navigateById(propertyValue, {
                 item: args.view.bindingContext,
+                context: instance.linkContext
+            });
+        });
+    }
+    else if (instance instanceof listViewModule.RadListView) {
+        instance.on("itemSelected", function (args) {
+            navigationModule.navigateById(propertyValue, {
+                item: instance.getSelectedItems()[0],
                 context: instance.linkContext
             });
         });
@@ -98,3 +107,4 @@ applicationModule.onLaunch = function (context) {
 };
 applicationModule.mainModule = viewsModule.initial;
 applicationModule.start();
+//# sourceMappingURL=app.js.map
