@@ -1,8 +1,3 @@
-var __extends = (this && this.__extends) || function (d, b) {
-    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-    function __() { this.constructor = d; }
-    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-};
 var typesModule = require("utils/types");
 var navigationModule = require("navigation");
 var constantsModule = require("../shared/constants");
@@ -232,3 +227,4 @@ exports.everlive = new everliveModule({
     },
 });
 exports.service = new Service();
+//# sourceMappingURL=service.js.map
