@@ -1,8 +1,3 @@
-var __extends = (this && this.__extends) || function (d, b) {
-    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-    function __() { this.constructor = d; }
-    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-};
 var view_model_base_1 = require("view-model-base");
 var service_1 = require("../../data/service");
 var validationRulesModule = require("validation-rules");
@@ -130,3 +125,4 @@ var EditProfileViewModel = (function (_super) {
     return EditProfileViewModel;
 })(view_model_base_1.ViewModelBase);
 exports.EditProfileViewModel = EditProfileViewModel;
+//# sourceMappingURL=edit-profile-view-model.js.map
