@@ -1,8 +1,3 @@
-var __extends = (this && this.__extends) || function (d, b) {
-    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-    function __() { this.constructor = d; }
-    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-};
 var edit_view_model_base_1 = require("edit-view-model-base");
 var service_1 = require("../../data/service");
 var view_report_view_model_1 = require("../view-report/view-report-view-model");
@@ -49,3 +44,4 @@ var EditReportViewModel = (function (_super) {
     return EditReportViewModel;
 })(edit_view_model_base_1.EditViewModelBase);
 exports.EditReportViewModel = EditReportViewModel;
+//# sourceMappingURL=edit-report-view-model.js.map
