@@ -1,8 +1,3 @@
-var __extends = (this && this.__extends) || function (d, b) {
-    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-    function __() { this.constructor = d; }
-    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-};
 var view_model_base_1 = require("view-model-base");
 var service_1 = require("../../data/service");
 var ViewExpenseViewModel = (function (_super) {
@@ -44,3 +39,4 @@ var ViewExpenseViewModel = (function (_super) {
     return ViewExpenseViewModel;
 })(view_model_base_1.ViewModelBase);
 exports.ViewExpenseViewModel = ViewExpenseViewModel;
+//# sourceMappingURL=view-expense-view-model.js.map
