@@ -3,7 +3,7 @@ var viewModel;
 function onNavigatingTo(args) {
     if (!args.isBackNavigation) {
         var page = args.object;
-        viewModel = new view_report_view_model_1.ViewReportViewModel(page.navigationContext.context.report);
+        viewModel = new view_report_view_model_1.ViewReportViewModel(page.navigationContext.item.report);
         page.bindingContext = viewModel;
     }
     viewModel.refresh();
@@ -13,3 +13,4 @@ function onItemTap(args) {
     viewModel.itemTap(args.view.bindingContext);
 }
 exports.onItemTap = onItemTap;
+//# sourceMappingURL=view-report.js.map
