@@ -5,7 +5,7 @@ declare module "application" {
     import cssSelector = require("ui/styling/css-selector");
     import observable = require("data/observable");
     import frame = require("ui/frame");
-
+    import {View} from "ui/core/view";
     /**
      * An extended JavaScript Error which will have the nativeError property initialized in case the error is caused by executing platform-specific code.
      */
@@ -77,6 +77,17 @@ declare module "application" {
     }
 
     /**
+     * Event data containing information for launch event.
+     */
+    export interface LaunchEventData extends ApplicationEventData {        
+        /**
+         * The root view for this Window on iOS or Activity for Android.
+         * If not set a new Frame will be created as a root view in order to maintain backwards compatibility.
+         */
+        root?: View;
+    }
+
+    /**
      * Event data containing information for orientation changed event.
      */
     export interface OrientationChangedEventData extends ApplicationEventData {
@@ -118,18 +129,19 @@ declare module "application" {
 
     /**
      * Loads css file and parses to a css syntax tree.
+     * @param cssFile Optional parameter to point to an arbitrary css file. If not specified, the cssFile property is used.
      */
-    export function loadCss(): void;
+    export function loadCss(cssFile?: string): Array<cssSelector.CssSelector>;
 
     /**
      * Call this method to start the application. Important: All code after this method call will not be executed!
      */
-    export function start();
+    export function start(entry?: frame.NavigationEntry);
 
     /**
      * The main entry point event. This method is expected to use the root frame to navigate to the main application page.
      */
-    export function onLaunch(context: any): void;
+    export function onLaunch(context?: any): void;
 
     /**
      * A callback to be used when an uncaught error occurs while the application is running.
@@ -173,7 +185,7 @@ declare module "application" {
      * @param callback - Callback function which will be removed.
      * @param thisArg - An optional parameter which will be used as `this` context for callback execution.
      */
-    export function off(eventNames: string, callback ?: any, thisArg ?: any);
+    export function off(eventNames: string, callback?: any, thisArg?: any);
 
     /**
      * Notifies all the registered listeners for the event provided in the data.eventName.
@@ -190,7 +202,7 @@ declare module "application" {
     /**
      * This event is raised on application launchEvent.
      */
-    export function on(event: "launch", callback: (args: ApplicationEventData) => void, thisArg?: any);
+    export function on(event: "launch", callback: (args: LaunchEventData) => void, thisArg?: any);
 
     /**
      * This event is raised when the Application is suspended.
@@ -331,11 +343,9 @@ declare module "application" {
         packageName: string;
 
         /**
-         * This method is called by the JavaScript Bridge when navigation to a new activity is triggered.
-         * @param intent - Native (android) intent used to create the activity.
-         * Returns com.tns.NativeScriptActivity.extend implementation.
+         * True if the application is not running (suspended), false otherwise.
          */
-        getActivity(intent: any /* android.content.Intent */): any;
+        paused: boolean;
 
         /**
          * [Deprecated. Please use the respective event instead.] Direct handler of the [onActivityCreated method](http://developer.android.com/reference/android/app/Application.ActivityLifecycleCallbacks.html).
