@@ -27073,12 +27073,12 @@ declare enum UITouchPhase {
 	UITouchPhaseCancelled
 }
 declare enum UIWebViewNavigationType {
-	UIWebViewNavigationTypeLinkClicked,
-	UIWebViewNavigationTypeFormSubmitted,
-	UIWebViewNavigationTypeBackForward,
-	UIWebViewNavigationTypeReload,
-	UIWebViewNavigationTypeFormResubmitted,
-	UIWebViewNavigationTypeOther
+	LinkClicked,
+	FormSubmitted,
+	BackForward,
+	Reload,
+	FormResubmitted,
+	Other
 }
 declare enum UIWebPaginationMode {
 	UIWebPaginationModeUnpaginated,
