@@ -1,5 +1,7 @@
 declare module "ui/enums" {
 
+    import animationModule = require("ui/animation");
+
     /**
      * Represents a soft keyboard flavor.
      */
@@ -9,25 +11,25 @@ declare module "ui/enums" {
          * iOS:  [UIKeyboardTypeNumbersAndPunctuation](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIKeyboardType)
          */
         export var datetime: string;
-        
+
         /**
          * Android: [TYPE_CLASS_PHONE](http://developer.android.com/reference/android/text/InputType.html#TYPE_CLASS_PHONE)
          * iOS:  [UIKeyboardTypePhonePad](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIKeyboardType)
          */
         export var phone: string;
-        
+
         /**
-         * Android: [TYPE_CLASS_NUMBER](http://developer.android.com/reference/android/text/InputType.html#TYPE_CLASS_NUMBER) | android.text.InputType.TYPE_NUMBER_VARIATION_NORMAL | [TYPE_NUMBER_FLAG_SIGNED](http://developer.android.com/reference/android/text/InputType.html#TYPE_NUMBER_FLAG_SIGNED) | [TYPE_NUMBER_FLAG_DECIMAL](http://developer.android.com/reference/android/text/InputType.html#TYPE_NUMBER_FLAG_DECIMAL)
+         * Android: [TYPE_CLASS_NUMBER](http://developer.android.com/reference/android/text/InputType.html#TYPE_CLASS_NUMBER) | [TYPE_NUMBER_VARIATION_NORMAL](http://developer.android.com/intl/es/reference/android/text/InputType.html#TYPE_NUMBER_VARIATION_NORMAL) | [TYPE_NUMBER_FLAG_SIGNED](http://developer.android.com/reference/android/text/InputType.html#TYPE_NUMBER_FLAG_SIGNED) | [TYPE_NUMBER_FLAG_DECIMAL](http://developer.android.com/reference/android/text/InputType.html#TYPE_NUMBER_FLAG_DECIMAL)
          * iOS:  [UIKeyboardTypeNumbersAndPunctuation](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIKeyboardType)
          */
         export var number: string;
-        
+
         /**
          * Android: [TYPE_CLASS_TEXT](http://developer.android.com/reference/android/text/InputType.html#TYPE_CLASS_TEXT) | [TYPE_TEXT_VARIATION_URI](http://developer.android.com/reference/android/text/InputType.html#TYPE_TEXT_VARIATION_URI)
          * iOS:  [UIKeyboardTypeURL](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIKeyboardType)
          */
         export var url: string;
-        
+
         /**
          * Android: [TYPE_CLASS_TEXT](http://developer.android.com/reference/android/text/InputType.html#TYPE_CLASS_TEXT) | [TYPE_TEXT_VARIATION_EMAIL_ADDRESS](http://developer.android.com/reference/android/text/InputType.html#TYPE_TEXT_VARIATION_EMAIL_ADDRESS)
          * iOS:  [UIKeyboardTypeEmailAddress](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIKeyboardType)
@@ -56,13 +58,13 @@ declare module "ui/enums" {
          * iOS: [UIReturnKeyGo](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIReturnKeyType)
          */
         export var go: string;
-        
+
         /**
          * Android: [IME_ACTION_SEARCH](http://developer.android.com/reference/android/view/inputmethod/EditorInfo.html#IME_ACTION_SEARCH)
          * iOS: [UIReturnKeySearch](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIReturnKeyType)
          */
         export var search: string;
-        
+
         /**
          * Android: [IME_ACTION_SEND](http://developer.android.com/reference/android/view/inputmethod/EditorInfo.html#IME_ACTION_SEND)
          * iOS: [UIReturnKeySend](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UITextInputTraits_Protocol/index.html#//apple_ref/c/tdef/UIReturnKeyType)
@@ -324,17 +326,17 @@ declare module "ui/enums" {
         export var none: string;
 
         /**
-         * Capitalize the first letter of each word automatically. 
+         * Capitalize the first letter of each word automatically.
          */
         export var words: string;
 
         /**
-         * Capitalize the first letter of each sentence automatically. 
+         * Capitalize the first letter of each sentence automatically.
          */
         export var sentences: string;
-        
+
         /**
-         * Capitalize all characters automatically. 
+         * Capitalize all characters automatically.
          */
         export var allCharacters: string;
     }
@@ -402,7 +404,7 @@ declare module "ui/enums" {
          */
         export var popup: string;
     }
-    
+
     /**
      * Specifies different font styles.
      */
@@ -411,7 +413,7 @@ declare module "ui/enums" {
          * Normal font style.
          */
         export var normal: string;
-        
+
         /**
          * Italic font style.
          */
@@ -426,7 +428,7 @@ declare module "ui/enums" {
          * No decoration.
          */
         export var none: string;
-        
+
         /**
          * Text decoration underline.
          */
@@ -446,7 +448,7 @@ declare module "ui/enums" {
          * No transform.
          */
         export var none: string;
-        
+
         /**
          * Text transform capitalize.
          */
@@ -456,7 +458,7 @@ declare module "ui/enums" {
          * Text transform uppercase.
          */
         export var uppercase: string;
-        
+
         /**
          * Text transform lowercase.
          */
@@ -471,13 +473,13 @@ declare module "ui/enums" {
          * Normal wrap.
          */
         export var normal: string;
-        
+
         /**
          * No wrap.
          */
         export var nowrap: string;
     }
-    
+
     /**
      * Specifies different font weights.
      */
@@ -486,13 +488,13 @@ declare module "ui/enums" {
          * Normal font weight.
          */
         export var normal: string;
-        
+
         /**
          * Bold font weight.
          */
         export var bold: string;
     }
-    
+
     /**
      * Specifies background repeat.
      */
@@ -522,6 +524,7 @@ declare module "ui/enums" {
      * Represents an animation curve type.
      */
     module AnimationCurve {
+
         /**
          * An ease-in curve causes the animation to begin slowly, and then speed up as it progresses.
          */
@@ -536,10 +539,20 @@ declare module "ui/enums" {
          * An ease-in ease-out curve causes the animation to begin slowly, accelerate through the middle of its duration, and then slow again before completing.
          */
         export var easeInOut: string;
-        
+
         /**
          * A linear animation curve causes an animation to occur evenly over its duration.
          */
         export var linear: string;
+
+        /**
+         * A spring animation curve causes an animation to produce a spring (bounce) effect.
+         */
+        export var spring: string;
+
+        /**
+         * A custom cubic bezier function defined by its two control points. Possible values are numeric values from 0 to 1
+         */
+        export function cubicBezier(x1: number, y1: number, x2: number, y2: number): animationModule.CubicBezierAnimationCurve;
     }
 }
